Creación del metodo DELETE
